#37 Move some duplicated code into abstract class
