+++ tests/phpunit/Civi/API/V4/Action/CreateWithOptionGroupTest.php
@@ -1,7 +1,6 @@
 <?php
-namespace Civi\API\V4;
+namespace Civi\API\V4\Action;
 
-use Civi\Api\TableDropperTrait;
 use Civi\Api4\CustomField;
 use Civi\Api4\CustomGroup;
 use Civi\Api4\Contact;
@@ -9,27 +8,7 @@ use Civi\Api4\Contact;
 /**
  * @group headless
  */
-class CreateWithOptionGroupTest extends UnitTestCase {
-
-  use TableDropperTrait;
-
-  /**
-   * Set up baseline for testing
-   */
-  public function setUp() {
-    $cleanup_params = array(
-      'tablesToTruncate' => array(
-        'civicrm_custom_group',
-        'civicrm_custom_field',
-        'civicrm_contact',
-        'civicrm_option_group',
-        'civicrm_option_value'
-      ),
-    );
-
-    $this->dropByPrefix('civicrm_value_mycontact');
-    $this->cleanup($cleanup_params);
-  }
+class CreateWithOptionGroupTest extends BaseCustomValueTest {
 
   public function testGetWithCustomData() {
     $customGroup = CustomGroup::create()
